refactor: global 하게 처리하지 않는 예외 처리 다른 클래스로 이동
