+++ src/main/java/roomescape/global/RoomescapeExceptionHandler.java
@@ -4,10 +4,6 @@ import java.util.NoSuchElementException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.reservation.domain.exception.PastReservationException;
-import roomescape.reservation.domain.exception.ReservationDateNullException;
-import roomescape.reservation.domain.exception.ReservationTimeNullException;
-import roomescape.reservation.domain.exception.ReserverNameEmptyException;
 
 @RestControllerAdvice
 public class RoomescapeExceptionHandler {
@@ -19,26 +15,6 @@ public class RoomescapeExceptionHandler {
 
     @ExceptionHandler(IllegalArgumentException.class)
     public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(PastReservationException.class)
-    public ResponseEntity<String> handelPastReservationException(PastReservationException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(ReserverNameEmptyException.class)
-    public ResponseEntity<String> handelReserverNameEmptyException(ReserverNameEmptyException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(ReservationDateNullException.class)
-    public ResponseEntity<String> handelReservationDateNullException(ReservationDateNullException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-
-    @ExceptionHandler(ReservationTimeNullException.class)
-    public ResponseEntity<String> handelReservationTimeNullException(ReservationTimeNullException e) {
         return ResponseEntity.badRequest().body(e.getMessage());
     }
 }
