+++ src/main/java/roomescape/reservation/controller/exception/ReservationExceptionHandler.java
@@ -4,6 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import roomescape.reservation.domain.exception.PastReservationException;
+import roomescape.reservation.domain.exception.ReservationDateNullException;
 import roomescape.reservation.domain.exception.ReserverNameEmptyException;
 
 @RestControllerAdvice
@@ -17,5 +18,10 @@ public class ReservationExceptionHandler {
     @ExceptionHandler(PastReservationException.class)
     public ResponseEntity<Void> handlePastReservationException(PastReservationException e) {
         return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(ReservationDateNullException.class)
+    public ResponseEntity<String> handelReservationDateNullException(ReservationDateNullException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
     }
 }
